+++ app/src/main/java/com/nighthawkapps/wallet/android/preference/Preferences.kt
@@ -1,0 +1,8 @@
+package com.nighthawkapps.wallet.android.preference
+
+import com.nighthawkapps.wallet.android.preference.model.BooleanDefaultValue
+
+object Preferences {
+    val isAcknowledgedAutoshieldingInformationPrompt =
+        BooleanDefaultValue(PreferenceKeys.IS_AUTOSHIELDING_INFO_ACKNOWLEDGED, false)
+}