+++ app/src/main/java/com/nighthawkapps/wallet/android/ui/MainActivity.kt
@@ -45,6 +45,7 @@ import androidx.core.content.getSystemService
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.NavController
+import androidx.navigation.NavDirections
 import androidx.navigation.Navigator
 import androidx.navigation.findNavController
 import cash.z.ecc.android.sdk.Initializer
@@ -74,7 +75,7 @@ import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-class MainActivity : AppCompatActivity() {
+class MainActivity : AppCompatActivity(R.layout.main_activity) {
 
     @Inject
     lateinit var clipboard: ClipboardManager
@@ -118,7 +119,6 @@ class MainActivity : AppCompatActivity() {
         }
         super.onCreate(savedInstanceState)
 
-        setContentView(R.layout.main_activity)
         initNavigation()
         initLoadScreen()
 
@@ -180,11 +180,13 @@ class MainActivity : AppCompatActivity() {
         navController?.popBackStack(destination, inclusive)
     }
 
-    fun safeNavigate(@IdRes destination: Int, extras: Navigator.Extras? = null) {
+    fun safeNavigate(navDirections: NavDirections) = safeNavigate(navDirections.actionId, navDirections.arguments, null)
+
+    fun safeNavigate(@IdRes destination: Int, args: Bundle? = null, extras: Navigator.Extras? = null) {
         if (navController == null) {
             navInitListeners.add {
                 try {
-                    navController?.navigate(destination, null, null, extras)
+                    navController?.navigate(destination, args, null, extras)
                 } catch (t: Throwable) {
                     twig(
                         "WARNING: during callback, did not navigate to destination: R.id.${
@@ -197,7 +199,7 @@ class MainActivity : AppCompatActivity() {
             }
         } else {
             try {
-                navController?.navigate(destination, null, null, extras)
+                navController?.navigate(destination, args, null, extras)
             } catch (t: Throwable) {
                 twig(
                     "WARNING: did not immediately navigate to destination: R.id.${