+++ buildSrc/src/main/java/com/nighthawkapps/wallet/android/Dependencies.kt
@@ -3,14 +3,15 @@ package com.nighthawkapps.wallet.android
 object Deps {
     // For use in the top-level build.gradle which gives an error when provided
     // `Deps.Kotlin.version` directly
-    const val kotlinVersion = "1.5.21"
-    const val navigationVersion =   "2.3.5"
-    const val compileSdkVersion = 30
+    const val kotlinVersion = "1.5.30"
+    const val navigationVersion = "2.3.5"
+    const val compileSdkVersion = 31
     const val buildToolsVersion = "31.0.0"
     const val minSdkVersion = 23
-    const val targetSdkVersion = 30
-    const val versionName = "1.0.26"
-    const val versionCode = 1_00_26_800 // last digits are alpha(0XX) beta(2XX) rc(4XX) release(8XX). Ex: 1_08_04_401 is an release candidate build of version 1.8.4 and 1_08_04_800 would be the final release.
+    const val targetSdkVersion = 31
+    const val versionName = "1.0.28"
+    const val versionCode =
+        1_00_28_800 // last digits are alpha(0XX) beta(2XX) rc(4XX) release(8XX). Ex: 1_08_04_401 is an release candidate build of version 1.8.4 and 1_08_04_800 would be the final release.
     const val packageName = "com.nighthawkapps.wallet.android"
 
     object AndroidX {
@@ -22,7 +23,7 @@ object Deps {
         const val FRAGMENT_KTX = "androidx.fragment:fragment-ktx:1.3.6"
         const val LEGACY = "androidx.legacy:legacy-support-v4:1.0.0"
         const val MULTIDEX = "androidx.multidex:multidex:2.0.1"
-        const val PAGING = "androidx.paging:paging-runtime-ktx:3.0.0"
+        const val PAGING = "androidx.paging:paging-runtime-ktx:3.0.1"
         const val RECYCLER = "androidx.recyclerview:recyclerview:1.2.0"
         const val SECURITY = "androidx.security:security-crypto:1.1.0-alpha03"
         const val DESUGAR_JDK = "com.android.tools:desugar_jdk_libs:1.1.5"
@@ -39,8 +40,8 @@ object Deps {
             }
         }
 
-        object Lifecycle : Version("2.2.0") {
-            val LIFECYCLE_RUNTIME_KTX = "androidx.lifecycle:lifecycle-runtime-ktx:2.4.0-alpha03"
+        object Lifecycle : Version("2.4.0-beta01") {
+            val LIFECYCLE_RUNTIME_KTX = "androidx.lifecycle:lifecycle-runtime-ktx:$version"
         }
 
         object Navigation : Version(navigationVersion) {
@@ -68,7 +69,7 @@ object Deps {
         const val GSON = "com.google.code.gson:gson:2.8.7"
     }
 
-    object Grpc : Version("1.40.0") {
+    object Grpc : Version("1.40.1") {
         val ANDROID = "io.grpc:grpc-android:$version"
         val OKHTTP = "io.grpc:grpc-okhttp:$version"
         val PROTOBUG = "io.grpc:grpc-protobuf-lite:$version"
@@ -82,8 +83,9 @@ object Deps {
 
     object Kotlin : Version(kotlinVersion) {
         val STDLIB = "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$version"
+        val REFLECT = "org.jetbrains.kotlin:kotlin-reflect:$version"
 
-        object Coroutines : Version("1.5.1") {
+        object Coroutines : Version("1.5.2") {
             val ANDROID = "org.jetbrains.kotlinx:kotlinx-coroutines-android:$version"
             val CORE = "org.jetbrains.kotlinx:kotlinx-coroutines-core:$version"
             val TEST = "org.jetbrains.kotlinx:kotlinx-coroutines-test:$version"
@@ -108,13 +110,18 @@ object Deps {
 
     object Test {
         const val JUNIT = "junit:junit:4.13.2"
-        const val MOKITO = "junit:junit:4.13.2"
+        const val MOKITO = "org.mockito:mockito-android:3.12.4"
         const val COROUTINES_TEST = "junit:junit:4.13.2"
         const val MOCKITO_KOTLIN = "com.nhaarman.mockitokotlin2:mockito-kotlin:2.2.0"
 
         object Android {
-            const val JUNIT = "androidx.test.ext:junit:1.1.1"
+            const val CORE = "androidx.test:core:1.4.0"
+            const val RULES = "androidx.test:rules:1.4.0"
+            const val JUNIT = "androidx.test.ext:junit:1.1.3"
+            const val FRAGMENT = "androidx.fragment:fragment-testing:1.4.0-alpha08"
             const val ESPRESSO = "androidx.test.espresso:espresso-core:3.4.0"
+            const val ESPRESSO_INTENTS = "androidx.test.espresso:espresso-intents:3.4.0"
+            const val NAVIGATION = "androidx.navigation:navigation-testing:2.3.0-alpha01"
         }
     }
 }